chore(EMS-2053): split up quote and application submission e2e tests workflow specs
