+++ e2e-tests/insurance/cypress/e2e/journeys/application-submission/multiple-policy-type/submit-an-application-multiple-policy-type-maximum-cover.spec.js
@@ -1,14 +1,15 @@
-import { APPLICATION } from '../../../../../constants';
-import { INSURANCE_ROUTES } from '../../../../../constants/routes/insurance';
+import { APPLICATION } from '../../../../../../constants';
+import { INSURANCE_ROUTES } from '../../../../../../constants/routes/insurance';
 
 const { APPLICATION_SUBMITTED } = INSURANCE_ROUTES;
 
-context('Insurance - submit an application - Multiple policy type, no broker - As an Exporter, I want to submit my completed export insurance application, So that UKEF can process and make a decision on my application', () => {
+context('Insurance - submit an application - Multiple policy type, no broker - As an Exporter, I want to submit my completed export insurance application with the maximum value of 500000, So that UKEF can process and make a decision on my application', () => {
   let referenceNumber;
 
   before(() => {
     cy.completeSignInAndSubmitAnApplication({
       policyType: APPLICATION.POLICY_TYPE.MULTIPLE,
+      policyAndExportsMaximumValue: true,
     }).then((refNumber) => {
       referenceNumber = refNumber;
     });
