+++ e2e-tests/insurance/cypress/e2e/journeys/application-submission/single-policy-type/submit-an-application-single-policy-type-different-contact-details.spec.js
@@ -1,4 +1,4 @@
-import { INSURANCE_ROUTES } from '../../../../../constants/routes/insurance';
+import { INSURANCE_ROUTES } from '../../../../../../constants/routes/insurance';
 
 const { APPLICATION_SUBMITTED } = INSURANCE_ROUTES;
